+++ mods/Worldwright/Data/Scripts/Worldwright/WorldwrightSession.cs
@@ -39,6 +39,7 @@ namespace Worldwright
         public override void UpdateAfterSimulation()
         {
             UpdateReclamationSpawnerEmissives();
+            UpdateReclamationSpawnerParticles();
             UpdatePendingReclamationSpawns();
         }